pyproject.toml
[project]
name = "apple-health-dashboard"
version = "0.1.0"
description = "Upload and explore Apple Health export data with a local dashboard."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Your Name" }]
dependencies = [
  "streamlit>=1.31",
  "pandas>=2.1",
  "altair>=5.0",
  "duckdb>=0.10.0",
  "polars>=0.20.0",
  "scikit-learn>=1.3",
  "statsmodels>=0.14",
  "google-generativeai>=0.3",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.2",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
